requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.1.7
openpyxl==3.1.5
gunicorn==21.2.0
